src/requirements.txt
pandas
google-cloud-alloydb-connector[pg8000]
google-cloud-storage
functions-framework
gradio
